move SentryConfig in src/model/config.ts
